Remove TinyMCE so it doesn't interfere with our boundaries
